refactor : package를 역할에 따라 service, controller 로 분리
LottoGameMain에서 LottoGame을 컨트롤러로 분리
LottoLogic을 Service 패키지로 이동
