+++ src/main/java/lotto/service/LottoLogic.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.service;
+
+import lotto.domain.*;
 
 import java.util.*;
