+++ src/test/java/lotto/service/LottoLogicTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.service;
 
+import lotto.domain.*;
+import lotto.service.LottoLogic;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
